Virtual Robot does not push
When using the console we want to be able to position the robots without moving blocks. When in concrete mode we want to have the virtual blocks update to reflect the concrete blocks. We accomplish this by having the virtual robot be a pusher while we are in concrete mode only.

 - split both robot and vrobot into pusher/non-pusher variants
 - when in console mode we use the non-pusher variants
 - when in concrete mode we use the pusher variants